Moved all DAG stuff into a separate "workflow" package and marked all the unsafe methods package-private
