+++ src/main/scala/pipelines/text/NewsgroupsPipeline.scala
@@ -10,6 +10,7 @@ import nodes.util.{Identity, CommonSparseFeatures, MaxClassifier}
 import org.apache.spark.{SparkConf, SparkContext}
 import pipelines.Logging
 import scopt.OptionParser
+import workflow.PipelineModel
 
 object NewsgroupsPipeline extends Logging {
   val appName = "NewsgroupsPipeline"
